refactor(android): share passkey login/signup token exchange
Both passkey login and signup finish with the same /oauth/token exchange via
Auth0.Android's signinWithPasskey, which accepts either an assertion or an
attestation as a JSON credential. Extract the identical exchange logic into a
PasskeyTokenExchangeApiRequestHandler base; PasskeyLoginApiRequestHandler and
PasskeySignupApiRequestHandler now only declare their method name.

Signup now uses the same assertHasProperties validation as the other handlers
(replacing bespoke null-checks) and drops its spurious API-28 guard — the token
exchange is a network call; only the example app's Credential Manager UI needs
API 28. Updated the signup handler test accordingly.
