+++ auth0_flutter/android/src/test/kotlin/com/auth0/auth0_flutter/request_handlers/api/PasskeySignupApiRequestHandlerTest.kt
@@ -12,6 +12,7 @@ import com.auth0.auth0_flutter.request_handlers.MethodCallRequest
 import io.flutter.plugin.common.MethodChannel.Result
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
+import org.junit.Assert.assertThrows
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.kotlin.*
@@ -42,7 +43,7 @@ class PasskeySignupApiRequestHandlerTest {
     )
 
     @Test
-    fun `should error when challenge is missing`() {
+    fun `should throw when challenge is missing`() {
         val options = hashMapOf<String, Any>("credential" to credentialMap())
         val handler = PasskeySignupApiRequestHandler()
         val mockApi = mock<AuthenticationAPIClient>()
@@ -50,13 +51,18 @@ class PasskeySignupApiRequestHandlerTest {
         val mockResult = mock<Result>()
         val request = MethodCallRequest(account = mockAccount, options)
 
-        handler.handle(mockApi, request, mockResult)
+        val exception = assertThrows(IllegalArgumentException::class.java) {
+            handler.handle(mockApi, request, mockResult)
+        }
 
-        verify(mockResult).error(eq("PASSKEY_ERROR"), eq("Missing challenge argument"), anyOrNull())
+        assertThat(
+            exception.message,
+            equalTo("Required property 'challenge.authSession' is not provided.")
+        )
     }
 
     @Test
-    fun `should error when credential is missing`() {
+    fun `should throw when credential is missing`() {
         val options = hashMapOf<String, Any>("challenge" to challengeMap())
         val handler = PasskeySignupApiRequestHandler()
         val mockApi = mock<AuthenticationAPIClient>()
@@ -64,9 +70,14 @@ class PasskeySignupApiRequestHandlerTest {
         val mockResult = mock<Result>()
         val request = MethodCallRequest(account = mockAccount, options)
 
-        handler.handle(mockApi, request, mockResult)
+        val exception = assertThrows(IllegalArgumentException::class.java) {
+            handler.handle(mockApi, request, mockResult)
+        }
 
-        verify(mockResult).error(eq("PASSKEY_ERROR"), eq("Missing credential argument"), anyOrNull())
+        assertThat(
+            exception.message,
+            equalTo("Required property 'credential' is not provided.")
+        )
     }
 
     @Test
